Fix setting tags for wrong glyphs
`textDidEndEditing:` is also send after switching from the edit view to the font view. The glyph selection has already changed to that of the font view but the tokens in the token field are still the tokens from the edit view. The did end message sets the tokens from the field to the selected glyphs and thus overwrites the tags for the glyphs selected in the font view. This listens instead for `textShouldEndEditing:` which happens before the glyph selection changes.